// frontend/src/HotelSearch/types.ts
export type Detail = {
	id: string;
	city: string;
	subCopy: string;
	price: number;
	image: string;
};

export type BookingCardProps = {
	id: string;
	city: string;
	subCopy: string;
	price: number;
	image: string;
	index: number;
};

export type HotelOption = {
	id: string;
	destination: string;
	image: string;
};